+++ app/projects/[project]/page.tsx
