+++ app/projects/page.tsx
@@ -1,9 +1,17 @@
 import Button from "@/components/button";
 import Hero from "@/components/hero";
+import CurrentProjects from "@/components/Projects/CurrentProjects";
+import PastProjects from "@/components/Projects/PastProjects";
+import projects from "./projectCards.json";
 
-export default function ProjectsPage() {
+export default function MainProjectPage() {
+  const currentYear = "2024-2025"
+  const currentProjects = projects[currentYear];
+  const pastProjects = Object.fromEntries(
+    Object.entries(projects).filter(([year]) => year !== currentYear)
+  );
   return (
-    <div>
+    <>
       <Hero
         title={<>Projects</>}
         subtitle={<>Check out our current and past projects!</>}
@@ -15,6 +23,9 @@ export default function ProjectsPage() {
           />
         }
       />
-    </div>
-  );
+      <CurrentProjects projects={currentProjects} />
+      <PastProjects projects={pastProjects} />
+    </>
+  )
 }
+
